Added state buffer boundary to prevent overruns.
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@837 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/src/GameSceneGraphComponent/SceneGraph.cpp
@@ -213,7 +213,7 @@ unsigned int SceneGraphManager::createGameEntity( const char *xmlText, int horde
 		{
 			SceneGraphComponent* sceneGraphComponent = new SceneGraphComponent( entity );
 			sceneGraphComponent->setHordeID( hordeID );
-			sceneGraphComponent->setMaxTrajection(atof(attachment.getAttribute("maxTrajection", "1.#INF")));
+			sceneGraphComponent->setMaxTrajection((float)atof(attachment.getAttribute("maxTrajection", "1.#INF")));
 
 #ifdef MEASURE_COMPONENT_LOADING
 			float time = GameEngine::currentTimeStamp();
